refactor: 拆分 Memory respond flow 长文件
